Define endpoint for project-version locales.
I chose to stick with "Iteration" when naming the classes, to keep
it consistent with the existing names. I do think we should change
the names, but doing it piecemeal will just lead to confusion and
will likely lead to a worse developer experience for an extended
time compared with just sticking with the old terminology.